requirements.txt
amqp==2.6.0
asgiref==3.2.7
beautifulsoup4==4.9.1
billiard==3.6.3.0
boto3==1.13.24
botocore==1.16.24
celery==4.4.5
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.0.6
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cleanup==4.0.0
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-storages==1.9.1
django-timezone-field==4.0
django-webpush==0.3.2
docutils==0.15.2
future==0.18.2
gunicorn==20.0.4
http-ece==1.1.0
idna==2.9
isort==4.3.21
jmespath==0.10.0
kombu==4.6.10
Pillow==7.1.2
psycopg2==2.8.5
py-vapid==1.7.0
pycparser==2.20
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
pywebpush==1.9.4
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
whitenoise==5.1.0